Migrate incidents control and fix missing values.
