+++ server_files/routes/routes.js
@@ -1,7 +1,9 @@
 import express from 'express';
 import sequelize from 'sequelize';
 import db from '../database/initDB.js';
-import { getAllIncidents } from '../controllers/incidents.js';
+import { createNewIncident, deleteIncident, getAllIncidents, getCallFromIncident, getDispatchFromIncident, getIncident, getLocationFromIncident, getUnitFromIncident, updateIncident } from '../controllers/incidents.js';
+
+
 import { getSearchResults } from '../controllers/search.js';
 
 const Op = sequelize.Op;
@@ -39,22 +41,7 @@ router.route('/incidents')
     await getAllIncidents(req, res);
   })
   .post(async (req, res) => {
-    try {
-      const newIncident = await db.incidents.create({
-        incident_id:  req.body.incident_id,
-        date: req.body.date,
-        description: req.body.description,
-        postal_code: req.body.postal_code,
-        district_code: req.body.district_code,
-        call_id: req.body.call_id,
-        dispatch_id: req.body.dispatch_id,
-        unit_id: req.body.unit_id
-      });
-      res.json(newIncident);
-    } catch (err) {
-      console.error(err);
-      res.send('Server Error!');
-    }
+    await createNewIncident(req, res);
   })
   .put(async (req, res) => {
     res.send('Action not available.');
@@ -65,170 +52,35 @@ router.route('/incidents')
 
 router.route('/incidents/:incident_id')
   .get(async (req, res) => {
-    try {
-      const incident = await db.incidents.findAll({
-        where: {
-          incident_id: req.params.incident_id
-        }
-      });
-      const reply = getReply(incident);
-      res.json(reply);
-    } catch (err) {
-      console.error(err);
-      res.send('Server Error!');
-    }
+    await getIncident(req, res);
   })
   .post(async (req, res) => {
     res.send('Action not available.');
   })
   .put(async (req, res) => {
-    try {
-      await db.incidents.update({
-        date: req.body.date,
-        description: req.body.description,
-        postal_code: req.body.postal_code,
-        district_code: req.body.district
-      },
-      {
-        where: {
-          incident_id: req.params.incident_id
-        }
-      });
-      res.send('Successful update.');
-    } catch (err) {
-      console.error(err);
-      res.send('Server Error!');
-    }
+    await updateIncident(req, res);
   })
   .delete(async (req, res) => {
-    try {
-      console.log(req.params)
-      const deleted = await db.incidents.destroy({
-        where: {
-          incident_id: req.params.incident_id
-        }
-      });
-      if (deleted > 0) {
-        res.send(`Deleted ${deleted} rows.`);
-      } else if (deleted === 0) {
-        res.send('No rows deleted.');
-      } else {
-        res.send('Server Error!');
-      }
-    } catch (err) {
-      console.error(err);
-      res.send('Server Error!');
-    }
+    await deleteIncident(req, res);
   });
 
 // Gets the unit from an incident
 router.get('/incidents/:incident_id/units', async (req, res) => {
-  try {
-    const getUnit = await db.incidents.findAll({
-      where: {
-        incident_id: req.params.incident_id
-      }
-    });    
-    console.log(getUnit);
-    
-    const allUnits = await db.units.findAll({
-      where: {
-        unit_id: getUnit[0].dataValues.unit_id
-      }
-    });
-
-    const reply = getReply(allUnits);
-    res.json(reply);
-  } catch (err) {
-    console.error(err);
-    res.send(err);
-  }
+  await getUnitFromIncident(req, res);
 });
 
 //Get calls from an incident
 router.get('/incidents/:incident_id/calls', async (req, res) => {
-  try {
-    const getIncidents = await db.incidents.findAll({
-      where: {
-        incident_id: req.params.incident_id
-      }
-    });    
-    
-    const allCalls = await db.calls.findAll({
-      where: {
-        call_id: getIncidents[0].dataValues.call_id
-      }
-    });
-
-    const reply = getReply(allCalls);
-    res.json(reply);
-  }
-   catch (err) {
-     console.error(err);
-     res.send(error);
-   }
+  await getCallFromIncident(req, res);
 });
 
 router.get('/incidents/:incident_id/dispatch', async (req, res) => {
-  try { 
-    const getIncidents = await db.incidents.findAll({
-      where: {
-        incident_id: req.params.incident_id
-      }
-    });    
-    const getDispatch = await db.dispatch.findAll({
-      where: {
-        dispatch_id: getIncidents[0].dataValues.dispatch_id
-      }
-    });
-    const reply = getReply(getDispatch);
-    res.json(reply);
-  }
-  catch (err) {
-    console.error(err);
-    res.send(error);
-  }
+  await getDispatchFromIncident(req, res);
 });
 
 router.get('/incidents/:incident_id/locations', async (req, res) => {
-  try { 
-    const getIncidents = await db.incidents.findAll({
-      where: {
-        incident_id: req.params.incident_id
-      
-      } 
-    });    
-    const getLocations = await db.locations.findAll({
-      where: {
-        locations_id: getIncidents[0].dataValues.locations_id
-      }
-    });
-    const reply = getReply(getLocations);
-    res.json(reply);
-  }
-  catch (err) {
-    console.error(err);
-    res.send(error);
-  }
+  await getLocationFromIncident(req, res);
 });
-
-// router.get('/incidents/:incident_id/incident', async (req, res) => {
-//   try { 
-//     const getIncidents = await db.incidents.findAll({
-//       where: {
-//         incident_id: req.params.incident_id
-//       }
-//     });    
-//     const reply = getReply(getIncidents);
-//     res.json(reply);
-//   }
-//   catch (err) {
-//     console.error(err);
-//     res.send(error);
-//   }
-// });
-
- 
 
 // CALLS
 router.route('/calls')
